Refactor RetryStrategy to be async
This refactors retry strategies to be async. This is needed for
when strategies need to internally wait or if they need to
protect access to a shared resource.
